pyproject.toml
[project]
authors = [
    {name = "Eugene Chintsov", email = "tr.e.chintsov@gmail.com"},
]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "pytest-cov>=6.0.0",
    "pytest>=8.3.4",
    "ruff>=0.8.4",
]

[project.scripts]
gendiff = "gendiff.scripts.gendiff_:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gendiff"]
